requirements.txt
Django==3.2.6
django_ckeditor==6.2.0
django_jalali==4.3.0
Pillow==8.4.0
python-decouple==3.5
requests==2.25.1
